requirements.txt
flask
requests
python-dotenv
google-generativeai
RateMyProfessorAPI